pyproject.toml
[project]
name = "rinha-fastapi"
version = "0.1.0"
description = "Rinha Backend 2025"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.116.1",
    "uvicorn[standard]>=0.32.0",
    "httpx>=0.28.1",
    "pydantic-settings>=2.10.1",
    "redis>=6.4.0",
    "uvloop>=0.19.0",
    "aio-pika>=9.5.7",
]

[tool.uv.workspace]
members = [
    "rinha",
]

[dependency-groups]
dev = [
    "black>=25.9.0",
    "isort>=6.0.1",
]
